+++ src/router/index.ts
@@ -3,5 +3,6 @@ export * from "./hash-router.js";
 export * from "./history-router.js";
 export * from "./memory-router.js";
 export * from "./path.js";
+export * from "./query.js";
 export * from "./route.js";
 export * from "./router.js";