+++ src/router/memory-router.ts
@@ -1,6 +1,7 @@
 import { batch, sig } from "../core/signals.js";
 import { normalize } from "./path.js";
-import { QueryInit, Router } from "./router.js";
+import { formatQuery, Query, QueryInit } from "./query.js";
+import { Router } from "./router.js";
 
 export interface MemoryRouterOptions {
 	/**
@@ -19,11 +20,11 @@ export interface MemoryRouterOptions {
  */
 export class MemoryRouter implements Router {
 	#path = sig<string>(undefined!);
-	#query = sig<URLSearchParams | undefined>(undefined);
+	#query = sig<Query | undefined>(undefined);
 
 	constructor(options?: MemoryRouterOptions) {
 		this.#path.value = normalize(options?.path ?? "");
-		this.#query.value = options?.query ? new URLSearchParams(options.query) : undefined;
+		this.#query.value = options?.query === undefined ? undefined : new Query(formatQuery(options.query));
 	}
 
 	get root(): Router {
@@ -38,14 +39,14 @@ export class MemoryRouter implements Router {
 		return this.#path.value;
 	}
 
-	get query(): URLSearchParams | undefined {
+	get query(): Query | undefined {
 		return this.#query.value;
 	}
 
 	push(path: string, query?: QueryInit): void {
 		batch(() => {
 			this.#path.value = normalize(path);
-			this.#query.value = query ? new URLSearchParams(query) : undefined;
+			this.#query.value = query === undefined ? undefined : new Query(formatQuery(query));
 		});
 	}